+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Version 1.2-alpha
+
+## Gameplay Changes
+increased target radius for player's auto-aim
+Fixed how the framerate works, now it works flawlessly on (hopefully) all browser.
+
+## Code Changes
+Removed all references to the level grid from entities, now it's all managed by the level itself.
+Created a new baseclass for level that doesn't use grid collision so I can compare performances. - The collision grid does in fact save A TON of performance.
+
 # Version 1.1-alpha
 
 ## New features